Basic schedule design, need to meet to clean architecture files and merge with add event view.

+++ src/main/java/entities/EventEntity/FlexibleEvent.java
@@ -22,7 +22,7 @@ public class FlexibleEvent implements Event {
     @JsonProperty("priorityLabel")
     private int priorityLabel;
     @JsonProperty("timeAllocation")
-    private int timeAllocation;
+    private float timeAllocation;
 
 
     public FlexibleEvent(LocalDateTime dayStart, LocalDateTime dayEnd, String eventName, int priorityLabel, float timeAllocation) {
